% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massDeFormula.R
\name{.atomicMasses}
\alias{.atomicMasses}
\title{Molecular mass for Elements}
\usage{
.atomicMasses()
}
\value{
This function returns a numeric matrix with mass values
}
\description{
This fuction returns the molecular mass based of main elements found in biology/proteomics as average and mono-isotopic mass. 
The result includes H, C, N, O, P, S, Se  and the electrone.
The values are bsed on http://www.ionsource.com/Card/Mass/mass.htm in ref to http://physics.nist.gov/Comp (as of 2019).
}
\examples{
.atomicMasses()
}
\seealso{
\code{\link{massDeFormula}}
}
