% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/razorNoFilter.R
\name{.checkKnitrProt}
\alias{.checkKnitrProt}
\title{Checking presence of knitr and rmarkdown}
\usage{
.checkKnitrProt(tryF = FALSE)
}
\arguments{
\item{tryF}{(logical)}
}
\value{
This function returns a logical value
}
\description{
This function allows checking presence of knitr and rmarkdown
}
\examples{
.checkKnitrProt()
}
\seealso{
\code{\link[wrMisc]{presenceFilt}}
}
