% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProteomeDiscovererFile.R
\name{.plotQuantDistr}
\alias{.plotQuantDistr}
\title{Generic plotting of density distribution for quantitation import-functions}
\usage{
.plotQuantDistr(
  abund,
  quant,
  custLay = NULL,
  normalizeMeth = NULL,
  softNa = NULL,
  refLi = NULL,
  refLiIni = NULL,
  notLogAbund = NA,
  figMarg = c(3.5, 3.5, 3, 1),
  tit = NULL,
  las = NULL,
  cexAxis = 0.8,
  nameSer = NULL,
  cexNameSer = NULL,
  silent = FALSE,
  callFrom = NULL,
  debug = FALSE
)
}
\arguments{
\item{abund}{(matrix or data.frame) abundance data, will be plottes as distribution}

\item{quant}{(matrix or data.frame) optional additional abundance data, to plot 2nd distribution, eg of normalized data}

\item{custLay}{(matrix) describing sammple-setup, typically produced by}

\item{normalizeMeth}{(character, length=1) name of normalization method (will be displayed in title of figure)}

\item{softNa}{(character, length=1) name of quantitation-software (typically 2-letter abbreviation, eg 'MQ')}

\item{refLi}{(integer) to display number reference lines}

\item{refLiIni}{(integer) to display initial number reference lines}

\item{notLogAbund}{(logical) set to \code{TRUE} if \code{abund} is linear but should be plotted as log2}

\item{figMarg}{(numeric, length=4) custom figure margins (will be passed to \code{\link[graphics]{par}}), defaults to c(3.5, 3.5, 3, 1)}

\item{tit}{(character) custom title}

\item{las}{(integer) indicate orientation of text in axes}

\item{cexAxis}{(numeric) size of numeric axis labels as cex-expansion factor (see also \code{\link[graphics]{par}})}

\item{nameSer}{(character) custom label for data-sets or columns (length must match number of data-sets)}

\item{cexNameSer}{(numeric) size of individual data-series labels as cex-expansion factor (see also \code{\link[graphics]{par}})}

\item{silent}{(logical) suppress messages}

\item{callFrom}{(character) allow easier tracking of messages produced}

\item{debug}{(logical) display additional messages for debugging}
}
\value{
This function returns logical value (if data were valid for plotting) and produces a density dustribution figure (if data were found valid)
}
\description{
This (low-level) function allows (generic) plotting of density distribution for quantitation import-functions
}
\examples{
set.seed(2018);  datT8 <- matrix(round(rnorm(800) +3,1), nc=8, dimnames=list(paste(
  "li",1:100,sep=""), paste(rep(LETTERS[1:3],c(3,3,2)),letters[18:25],sep="")))
.plotQuantDistr(datT8, quant=NULL, refLi=NULL, tit="Synthetic Data Distribution")                                
}
\seealso{
used in \code{readProtDiscovererFile},  \code{\link{readMaxQuantFile}}, \code{\link{readProlineFile}}, \code{\link{readFragpipeFile}}
}
