% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUPS1acc.R
\name{getUPS1acc}
\alias{getUPS1acc}
\title{Accession-Numbers And Names Of UPS1 Proteins}
\usage{
getUPS1acc()
}
\value{
This function returns data.frame with accession-numbers as stated by the supplier (\code{$acFull}),
 trimmed accession-numbers, ie without version numbers (\code{$ac}) 
 and associated (\code{UniProt}) names on \href{https://www.uniprot.org/}{UniProt} as well as the species designation for the collection of 48 human UPS-1 proteins.
}
\description{
UPS1 (see https://www.sigmaaldrich.com/FR/en/product/sigma/ups1) and UPS2 are commerical products consisting of a mix of 48 human (purified) proteins.
They are frequently used as standard in spike-in experiments, available from \href{https://www.sigmaaldrich.com}{Sigma-Aldrich}.
This function allows accessing their protein accession numbers and associated names on \href{https://www.uniprot.org/}{UniProt}
}
\examples{
head(getUPS1acc())
}
