% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectSpeciesIndic.R
\name{inspectSpeciesIndic}
\alias{inspectSpeciesIndic}
\title{Inspect Species Indictaion Or Group of Proteins}
\usage{
inspectSpeciesIndic(x, silent = FALSE, debug = FALSE, callFrom = NULL)
}
\arguments{
\item{x}{(character) species indication or name of collection of proteins (so far only UPS1 & UPS2)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) display additional messages for debugging}

\item{callFrom}{(character) allows easier tracking of messages produced}
}
\value{
This function returns a character vector
}
\description{
This function inspects its main argument to convert a species indication to the scientific name or to return all protein-accession numbers for a name of a standard collection like UPS1.
}
\examples{
inspectSpeciesIndic("Human")
inspectSpeciesIndic("UPS1")
}
\seealso{
\code{\link{getUPS1acc}};
}
