% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massDeFormula.R
\name{massDeFormula}
\alias{massDeFormula}
\title{Molecular mass from chemical formula}
\usage{
massDeFormula(
  comp,
  massTy = "mono",
  rmEmpty = FALSE,
  silent = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{comp}{(character) atomic compostion}

\item{massTy}{(character) 'mono' or 'average'}

\item{rmEmpty}{(logical) suppress empty entries}

\item{silent}{(logical) suppress messages}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a numeric vector with mass
}
\description{
Calculate molecular mass based on atomic composition
}
\examples{
massDeFormula(c("12H12O","HO"," 2H 1 Se, 6C 2N","HSeCN"," ","e"))
}
\seealso{
\code{\link[wrMisc]{convToNum}}
}
