% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProteomeDiscovererPeptides.R
\name{readProtDiscovererPeptides}
\alias{readProtDiscovererPeptides}
\title{readProtDiscovererPeptides, depreciated}
\usage{
readProtDiscovererPeptides(...)
}
\arguments{
\item{...}{Actually, this function doesn't ready any input any more}
}
\value{
This function returns \code{NULL}
}
\description{
This function has been depreciated and replaced by \code{\link{readProteomeDiscovererPeptides}} (from this package).
}
\seealso{
\code{\link{readProteomeDiscovererFile}}, \code{\link{readProteomeDiscovererPeptides}}
}
