% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortSoftwName.R
\name{shortSoftwName}
\alias{shortSoftwName}
\title{Get Short Names of Proteomics Quantitation Software}
\usage{
shortSoftwName(
  x,
  tryAsLower = TRUE,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{x}{(character) 'mono' or 'average'}

\item{tryAsLower}{(logical)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a vector with masses for all amino-acids (argument 'massTy' to switch from mono-isotopic to average mass)
}
\description{
Get/convert short names of various proteomics quantitation software names.
A 2-letter abbreviation will be returned
}
\examples{
shortSoftwName(c("maxquant","DIANN"))
}
\seealso{
\code{\link{massDeFormula}}, \code{\link[wrMisc]{convToNum}}
}
