% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProtDiscovFile.R
\name{.commonSpecies}
\alias{.commonSpecies}
\title{Get matrix with UniProt abbreviations for common species}
\usage{
.commonSpecies()
}
\value{
This function returns a 2-column matrix with species names
}
\description{
This (low-level) function allows accessing matrix with UniProt abbreviations for common species
This information maybe used to harmonize species descriptions.
}
\examples{
.commonSpecies()
}
\seealso{
used eg in \code{readProtDiscovererFile},  \code{\link{readMaxQuantFile}}, \code{\link{readProlineFile}}, \code{\link{readFragpipeFile}}
}
