% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUPS1acc.R
\name{getUPS1acc}
\alias{getUPS1acc}
\title{Accession-Numbers And Names Of UPS-1 Proteins}
\usage{
getUPS1acc()
}
\value{
This function returns data.frame with accession-numbers as stated by the supplier (\code{$acFull}),
 trimmed accession-numbers ie without version numbers (\code{$ac}) 
 and associated Names (\code{uniProt}) on href{https://www.uniprot.org/}{UniProt} as well as the species designation for the collection of 48 human UPS-1 proteins.
}
\description{
\href{https://www.sigmaaldrich.com/FR/en/product/sigma/ups1}{UPS-1} and UPS-2 are mix of 48 of human proteins frequently used 
as standard in spike-in experiments. They are comercially available from Sigma-Aldrich.
This function allows accessing their accession-numbers and associated Names on href{https://www.uniprot.org/}{UniProt}
}
\examples{
head(getUPS1acc())
}
