% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortSoftwName.R
\name{shortSoftwName}
\alias{shortSoftwName}
\title{Get Short Names of Proteomics Quantitation Software}
\usage{
shortSoftwName(
  x,
  tryAsLower = TRUE,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{x}{(character) software (full) name}

\item{tryAsLower}{(logical) include lower-caps writing to search}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a vector with 2-letter abbreviation for the software
}
\description{
Get/convert short names of various proteomics quantitation software names for software results handeled by this package.
A 2-letter abbreviation will be returned
}
\details{
So far thuis function recognizes the following software names:
"DIA-NN", "ProteomeDiscoverer", "Compomics", "MaxQuant", "Proline", "TPP", "FragPipe", "MassChroQ", "OpenMS", "Ionbot" and "Sage"
}
\examples{
shortSoftwName(c("maxquant","DIANN"))
}
\seealso{
\code{\link{readMaxQuantFile}}
}
